+++ src/tests/TestConstants.ts
@@ -10,4 +10,9 @@ export class TestConstants
      * Gets the path to a folder for running the test-workspace.
      */
     public static readonly TestWorkspaceDirectory = join(Constants.PackageDirectory, "test");
+
+    /**
+     * Gets the name of an event for testing.
+     */
+    public static readonly TestEvent = "typingsInstallerPid";
 }