+++ src/tests/TestLanguageServiceTester.ts
@@ -7,11 +7,6 @@ import { TestConstants } from "./TestConstants";
 export class TestLanguageServiceTester extends LanguageServiceTester
 {
     /**
-     * The error-codes to test.
-     */
-    private errorCodes: number[] = [];
-
-    /**
      * Initializes a new instance of the {@link TestLanguageServiceTester `TestLanguageServiceTester`} class.
      *
      * @param workingDirectory
@@ -20,21 +15,5 @@ export class TestLanguageServiceTester extends LanguageServiceTester
     public constructor(workingDirectory = TestConstants.TestWorkspaceDirectory)
     {
         super(workingDirectory);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public get ErrorCodes(): number[]
-    {
-        return this.errorCodes;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public set ErrorCodes(value)
-    {
-        this.errorCodes = value;
     }
 }